backend/requirements.txt
fastapi[all]
pydantic
pydantic-settings
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
faiss-cpu
sentence-transformers
python-multipart
python-docx
docx2txt
pdfplumber
pypdf
supabase
python-dotenv
uvicorn
